# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/file-cid/{id}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "取得指定 File",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "FileCID"
                ],
                "summary": "取得指定 File",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "FileCID ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/api/file-cid/{name}": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    }
                ],
                "description": "上傳 File",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FileCID"
                ],
                "summary": "上傳 File",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.FileCID"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.FileCID": {
            "type": "object",
            "properties": {
                "cid": {
                    "description": "cid of ipfs",
                    "type": "string",
                    "example": "QmWMgCrvNEoD6KnqFfnv4tz7X6soDJvioswNBjHt6XcEum"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "name",
                    "type": "string",
                    "example": "meeting.txt"
                }
            }
        }
    }
}